2319: Move shared functionality and components to `/src/shared`
