+++ administration/src/bp-modules/applications/ApplicationCard.tsx
@@ -49,14 +49,14 @@ import {
 import BaseMenu, { MenuItemType } from '../../mui-modules/base/BaseMenu'
 import { ProjectConfigContext } from '../../project-configs/ProjectConfigContext'
 import type { ProjectConfig } from '../../project-configs/getProjectConfig'
+import JsonFieldView from '../../shared/components/JsonFieldView'
+import VerificationsView from '../../shared/components/VerificationsView'
 import { ApplicationDataIncompleteError } from '../../util/applicationDataHelper'
 import formatDateWithTimezone from '../../util/formatDate'
 import getApiBaseUrl from '../../util/getApiBaseUrl'
 import { useAppToaster } from '../AppToaster'
 import { AccordionExpandButton } from '../components/AccordionExpandButton'
-import JsonFieldView from './JsonFieldView'
 import NoteDialogController from './NoteDialogController'
-import VerificationsView from './VerificationsView'
 import { ApplicationIndicators } from './components/VerificationsIndicator'
 import type { Application } from './types'
 import { ApplicationToCsvError, exportApplicationToCsv } from './utils/exportApplicationToCsv'
