+++ administration/src/bp-modules/applications/ApplicationStatusBar.test.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 
-import { ApplicationStatus } from '../../../generated/graphql'
-import { renderWithTranslation } from '../../../testing/render'
-import ApplicationStatusBar from '../ApplicationStatusBar'
-import { barItems } from '../ApplicationsOverview'
-import type { Application } from '../types'
+import { ApplicationStatus } from '../../generated/graphql'
+import { renderWithTranslation } from '../../testing/render'
+import ApplicationStatusBar from './ApplicationStatusBar'
+import { barItems } from './ApplicationsOverview'
+import type { Application } from './types'
 
 const applications: Application[] = [
   {
