+++ administration/src/bp-modules/applications/__mocks__/applicationData.ts
@@ -1,5 +1,5 @@
 import { ApplicationStatus } from '../../../generated/graphql'
-import { JsonField } from '../JsonFieldView'
+import { JsonField } from '../../../shared/components/JsonFieldView'
 import type { Application } from '../types'
 import { verificationsAwaiting } from './verificationData'
