+++ administration/src/bp-modules/applications/constants/index.ts
@@ -1,8 +1,1 @@
-import { css } from 'styled-components'
-
-export const printAwareCss = css`
-  @media print {
-    display: none;
-  }
-`
 export const CSV_MIME_TYPE_UTF8 = 'text/csv;charset=utf-8;'
