--- administration/src/bp-modules/applications/utils/index.ts
@@ -1,13 +1,0 @@
-import { type ApplicationVerification, VerificationStatus } from '../types'
-
-export const verificationStatus = (
-  verification: Pick<ApplicationVerification, 'rejectedDate' | 'verifiedDate'>
-): VerificationStatus => {
-  if (verification.verifiedDate) {
-    return VerificationStatus.Verified
-  }
-  if (verification.rejectedDate) {
-    return VerificationStatus.Rejected
-  }
-  return VerificationStatus.Pending
-}
