--- administration/src/bp-modules/applications/utils/verificationHelper.ts
@@ -1,1 +1,0 @@
-export const isEmailValid = (email: string): boolean => /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email)
