+++ administration/src/mui-modules/application/primitive-inputs/EmailForm.tsx
@@ -1,9 +1,9 @@
 import { TextField } from '@mui/material'
 import React, { useContext, useState } from 'react'
 
-import { isEmailValid } from '../../../bp-modules/applications/utils/verificationHelper'
 import { EmailInput } from '../../../generated/graphql'
 import i18next from '../../../i18n'
+import { isEmailValid } from '../../../shared/verifications'
 import FormAlert from '../../base/FormAlert'
 import { FormContext } from '../SteppedSubForms'
 import { Form, FormComponentProps } from '../util/FormType'
