+++ administration/src/shared/components/JsonFieldView.tsx
